+++ src/Bisarca/Graph/Edge/DirectedEdgeInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the bisarca/graph package.
+ *
+ * (c) Emanuele Minotto <minottoemanuele@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Bisarca\Graph\Edge;
+
+/**
+ * Interface used to implement a directed edges.
+ */
+interface DirectedEdgeInterface extends EdgeInterface
+{
+}