+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All Notable changes to `bisarca/graph` will be documented in this file.
 ### Added
 - edges, vertices, graph and sets identification using `Bisarca\Graph\Identifier\IdentifierAwareInterface` and `Bisarca\Graph\Identifier\IdentifierAwareTrait`
 - edges documentation
+- added directed edges with interface
 
 ### Deprecated
 - Nothing